Uprava workflow hlaseni.

+++ trunk/BIStro/reports.php
@@ -85,6 +85,7 @@
 	        ".DB_PREFIX."reports.datum AS 'datum',
 	        ".DB_PREFIX."reports.label AS 'label',
 	        ".DB_PREFIX."reports.task AS 'task',
+	        ".DB_PREFIX."reports.status AS 'status',
 	        ".DB_PREFIX."users.login AS 'autor',
 	        ".DB_PREFIX."reports.iduser AS 'iduser',
 	        ".DB_PREFIX."reports.type AS 'type'
@@ -96,7 +97,7 @@
 	while ($rec=MySQL_Fetch_Assoc($res)) {
 	  echo '<div class="news_div '.(($rec['type']==1)?'game_news':'system_news').'">
 	<div class="news_head"><strong><a href="readactrep.php?rid='.$rec['id'].'">'.StripSlashes($rec['label']).'</a></strong>';
-	  if (($usrinfo['right_power']) || $usrinfo['id']==$rec['iduser']) {
+	  if (($usrinfo['right_power']) || ($usrinfo['id']==$rec['iduser'] && $rec['status']<1)) {
 	   	echo '	 | <td><a href="editactrep.php?rid='.$rec['id'].'">upravit</a> | <a href="procactrep.php?delete='.$rec['id'].'" onclick="'."return confirm('Opravdu smazat hlášení &quot;".StripSlashes($rec['label'])."&quot;?');".'">smazat</a></td>';
 	  }
 	  echo '.</span>
